Refactor: 스터디룸 생성 함수 위치 변경
