+++ src/pages/study-rooms/components/create-study-room-modal.tsx
@@ -1,11 +1,17 @@
+import { useNavigate } from 'react-router-dom';
 import { TextField } from '@mui/material';
 import { MODAL_KEYS } from '@/constant/modal';
+import { ROUTE_PATH } from '@/constant/routes';
+import { createStudyRoom as createStudyRoomApi } from '@/apis/study-room/create-study-room';
 import useModal from '@/hooks/useModal';
 import useCreateStudyRoomForm from '../hooks/useCreateStudyRoomForm';
 import Modal from '@/components/common/modal/modal';
 import CreateStudyRoomSelects from './create-study-room-select-fields';
+import { setLocalStorage } from '@/utils/storage';
 
 const CreateStudyRoomModal = () => {
+  const navigate = useNavigate();
+
   const { isOpen, onClose } = useModal(MODAL_KEYS.CREATE_STUDY_ROOM);
 
   const {
@@ -13,11 +19,20 @@ const CreateStudyRoomModal = () => {
     timeSet,
     handleChangeTextField,
     handleChangeSelect,
-    createStudyRoom,
   } = useCreateStudyRoomForm();
 
-  const handleClickCreateStudyRoom = () => {
-    createStudyRoom({ name, intro, timeSet });
+  const createStudyRoom = async () => {
+    try {
+      const { id, participantId } = await createStudyRoomApi({ name, intro, timeSet });
+
+      setLocalStorage({ key: 'participantId', value: String(participantId) });
+
+      onClose();
+
+      navigate(`${ROUTE_PATH.STUDY_ROOMS}/${id}`);
+    } catch (e) {
+      console.error(e);
+    }
   };
 
   return (
@@ -26,7 +41,7 @@ const CreateStudyRoomModal = () => {
       title="스터디룸 생성"
       closeBtn="취소"
       actionBtn="방 만들기"
-      onClickActionBtn={handleClickCreateStudyRoom}
+      onClickActionBtn={createStudyRoom}
       onClose={onClose}
     >
       <TextField
